fix: concurrent event handler map writes when services start up
